[WIP] Started working on server architecture with player.status.

+++ game/Weapon.cpp
@@ -26,7 +26,7 @@ Weapon::Weapon(const Weapon& other)
 
 // ===== GETTERS =====
 
-int Weapon::getId() const {
+short Weapon::getId() const {
     return w_id;
 }
